Domain adaptation moved to a seprate file
